THIS IS THE BEST SONG EVER MADE IN THE WORLD
This is the best song ever made in the world
This is the best song ever made in the world
Other songs are good, even some are really good
But they are not the best songs ever made in the world

It has a lot of good beats and a lot of good singing
And the music is so good, even scientists don't know what it is
You can listen to it anywhere, on a bike or on a different bike
You can dance to it any time, in the morning or with your arms
You can also dance to it at other times, like at lunch
What are you gonna have for lunch? I'm gonna have two soups
If you already haven't noticed, soup is my favorite food
I also really like my watch, and my favorite song is this song

This is the best song ever made in the world
This is the best song ever made in the world
Other songs make people dance, like Beyonce and Eminem's
But they are not the best songs ever made in the world

If you like listening to songs with your ears, well this is the song
for you
You can listen to it while you jog or while you're running away from
cars
My friend who is a mail man even likes this song
I guess a lot of mail gets delivered here, 'cause he stands on this
corner a lot
If you like eating sandwiches, you'll probably like this song
Because just like sandwiches, this song is not made out of glass
A glass sandwich, that would be gross, I'll just stick with soup
Did I mention soup's my favorite food? ...Oh yeah, I did

This is the best song ever made in the world
This is the best song ever made in the world
Other songs have different instruments, like a guitar or a girl
But they are not the best songs ever made in the world

If you like this song a lot, say that you like it a lot
I like it a lot
And if you think this song is the best, have a bowl of soup with me
[slurping] Oh boy, that is good soup
And if you think this song is the best in the world, have a second bowl
of soup with me
[slurping] That is also good soup
Ha ha ha ha ha ha, I tricked you, I don't even like soup
But you thought I did, ha ha ha ha, oh boy
I'm so good at tricking people into thinking that I like soup
Ever since my whole family got drowned, it's my favorite thing to do

This is the best song ever made in the world
This is the best song ever made in the world
Also my best friend got killed when he got hit with a snake
But this is the best song ever made in the world